fix: Use handshake version 2 (#3157)
* fix: Use both version and oldest_supported version

This will allow nodes that are compatible with each other to
communicate disregarding who initiated the conversation.

Test plan
=========
Pass upgradability test

* Fix failing test

* fix: Use handshake version 2

**DO NOT MERGE**

This change is only backward compatible with branch
latest_supported_version. We must merge and deploy that branch first.

Test plan
=========

1. Check latest_supported_version_2 branch is backward compatible with
latest_supported_version branch.

Run backward.py with
main(near_root, "latest_supported_version", "latest_supported_version_2")

2. handshake.py pass

* Update handshake.py

* Use previous version as default version for incoming handshakes

* Use +1 predicted version

* Respond with current version to other nodes

* Reaction to ProtocolVersionMismatch and better API

* Update backward compatible test

* Remove Option

* Fix codec test

* Backward compatible with version bump

* hotfix upgradable tests

* Address bowen comments

* test fix

* test another attempt

* compile current

* add debug info

* try again

* remove debug info

* Remove unnecessary version checks

* Add responses to Handshake with future version

* Fix test

* Don't use match

+++ pytest/lib/branches.py
@@ -38,7 +38,10 @@ def compile_current():
         subprocess.check_output(['cargo', 'build', '-p', 'near'])
     subprocess.check_output(['cargo', 'build', '-p', 'state-viewer'])
     branch = escaped(branch)
-    os.rename('../target/debug/near', '../target/debug/near-%s' % branch)
+    if os.path.exists('../target/debug/near'):
+        os.rename('../target/debug/near', '../target/debug/near-%s' % branch)
+    else:
+        os.rename('../target/debug/neard', '../target/debug/near-%s' % branch)
     os.rename('../target/debug/state-viewer',
               '../target/debug/state-viewer-%s' % branch)
     subprocess.check_output(['git', 'checkout', '../Cargo.lock'])
@@ -64,11 +67,17 @@ def download_binary(uname, branch):
 
 def prepare_ab_test(other_branch):
     # Use NEAR_AB_BINARY_EXISTS to avoid rebuild / re-download when testing locally.
-    if not os.environ.get('NEAR_AB_BINARY_EXISTS'):
-        compile_current()
-        uname = os.uname()[0]
-        if other_branch in ['master', 'beta', 'stable'] and uname in ['Linux', 'Darwin']:
-            download_binary(uname, other_branch)
-        else:
-            compile_binary(other_branch)
+    #if not os.environ.get('NEAR_AB_BINARY_EXISTS'):
+    #    compile_current()
+    #    uname = os.uname()[0]
+    #    if other_branch in ['master', 'beta', 'stable'] and uname in ['Linux', 'Darwin']:
+    #        download_binary(uname, other_branch)
+    #    else:
+    # TODO: re-enable caching
+    compile_current()
+    uname = os.uname()[0]
+    try:
+        download_binary(uname, other_branch)
+    except Exception:
+        compile_binary(other_branch)
     return '../target/debug/', [other_branch, escaped(current_branch())]
